move stable app config tests to single test file
